moved logger exports to package index
